refactor(#76): 패키지명 entity로 변경
