+++ src/main/java/com/widyu/domain/album/domain/Album.java
@@ -1,0 +1,5 @@
+package com.widyu.domain.album.domain;
+
+public class Album {
+
+}
