+++ src/main/java/com/widyu/domain/auth/application/guardian/GuardianAuthService.java
@@ -4,7 +4,7 @@ import com.widyu.domain.auth.application.LogoutService;
 import com.widyu.domain.auth.application.TemporaryTokenService;
 import com.widyu.domain.auth.application.guardian.local.LocalLoginService;
 import com.widyu.domain.auth.application.guardian.oauth.SocialLoginService;
-import com.widyu.domain.auth.domain.TemporaryMember;
+import com.widyu.domain.auth.entity.TemporaryMember;
 import com.widyu.domain.auth.dto.TemporaryTokenDto;
 import com.widyu.domain.auth.dto.request.AppleSignUpRequest;
 import com.widyu.domain.auth.dto.request.ChangePasswordRequest;
@@ -26,8 +26,8 @@ import com.widyu.global.util.MemberUtil;
 import com.widyu.domain.auth.dto.response.TemporaryTokenResponse;
 import com.widyu.domain.auth.dto.response.TokenPairResponse;
 import com.widyu.domain.auth.dto.response.UserProfile;
-import com.widyu.domain.member.domain.Member;
-import com.widyu.domain.member.domain.SocialAccount;
+import com.widyu.domain.member.entity.Member;
+import com.widyu.domain.member.entity.SocialAccount;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
