+++ src/main/java/com/widyu/domain/auth/application/guardian/GuardianTokenService.java
@@ -5,8 +5,8 @@ import com.widyu.domain.auth.dto.request.RefreshTokenRequest;
 import com.widyu.domain.auth.dto.response.TokenPairResponse;
 import com.widyu.global.security.JwtTokenProvider;
 import com.widyu.global.util.MemberUtil;
-import com.widyu.domain.member.domain.Member;
-import com.widyu.domain.member.domain.MemberRole;
+import com.widyu.domain.member.entity.Member;
+import com.widyu.domain.member.entity.MemberRole;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
