+++ src/main/java/com/widyu/domain/auth/application/guardian/oauth/SocialLoginService.java
@@ -3,8 +3,8 @@ package com.widyu.domain.auth.application.guardian.oauth;
 import com.widyu.domain.auth.application.guardian.oauth.strategy.SocialLoginStrategy;
 import com.widyu.domain.auth.application.guardian.oauth.strategy.SocialLoginStrategyFactory;
 import com.widyu.domain.auth.application.guardian.oauth.strategy.UserInfo;
-import com.widyu.domain.auth.domain.OAuthProvider;
-import com.widyu.domain.auth.domain.TemporaryMember;
+import com.widyu.domain.auth.entity.OAuthProvider;
+import com.widyu.domain.auth.entity.TemporaryMember;
 import com.widyu.domain.auth.dto.request.AppleSignUpRequest;
 import com.widyu.domain.auth.dto.request.SocialIntegrationRequest;
 import com.widyu.domain.auth.dto.request.SocialLoginRequest;
@@ -17,10 +17,10 @@ import com.widyu.global.error.BusinessException;
 import com.widyu.global.error.ErrorCode;
 import com.widyu.global.security.JwtTokenProvider;
 import com.widyu.global.util.TemporaryMemberUtil;
-import com.widyu.domain.member.domain.Member;
-import com.widyu.domain.member.domain.MemberRole;
-import com.widyu.domain.member.domain.MemberType;
-import com.widyu.domain.member.domain.SocialAccount;
+import com.widyu.domain.member.entity.Member;
+import com.widyu.domain.member.entity.MemberRole;
+import com.widyu.domain.member.entity.MemberType;
+import com.widyu.domain.member.entity.SocialAccount;
 import com.widyu.domain.member.repository.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
