+++ src/main/java/com/widyu/domain/auth/application/parent/ParentAuthService.java
@@ -7,9 +7,9 @@ import com.widyu.global.error.BusinessException;
 import com.widyu.global.error.ErrorCode;
 import com.widyu.global.security.JwtTokenProvider;
 import com.widyu.global.util.MemberUtil;
-import com.widyu.domain.member.domain.Member;
-import com.widyu.domain.member.domain.MemberType;
-import com.widyu.domain.member.domain.ParentProfile;
+import com.widyu.domain.member.entity.Member;
+import com.widyu.domain.member.entity.MemberType;
+import com.widyu.domain.member.entity.ParentProfile;
 import com.widyu.domain.member.repository.MemberRepository;
 import com.widyu.domain.member.repository.ParentProfileRepository;
 import java.util.ArrayList;
