+++ src/main/java/com/widyu/domain/auth/dto/TemporaryTokenDto.java
@@ -1,6 +1,6 @@
 package com.widyu.domain.auth.dto;
 
-import com.widyu.domain.member.domain.MemberRole;
+import com.widyu.domain.member.entity.MemberRole;
 
 public record TemporaryTokenDto(
         String temporaryMemberId,
