+++ src/main/java/com/widyu/domain/auth/entity/TokenType.java
@@ -1,4 +1,4 @@
-package com.widyu.domain.auth.domain;
+package com.widyu.domain.auth.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
