+++ src/main/java/com/widyu/domain/auth/repository/OAuthStateRepository.java
@@ -1,6 +1,6 @@
 package com.widyu.domain.auth.repository;
 
-import com.widyu.domain.auth.domain.OAuthState;
+import com.widyu.domain.auth.entity.OAuthState;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
