+++ src/main/java/com/widyu/domain/auth/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package com.widyu.domain.auth.repository;
 
-import com.widyu.domain.auth.domain.RefreshToken;
+import com.widyu.domain.auth.entity.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
