+++ src/main/java/com/widyu/domain/auth/repository/TemporaryMemberRepository.java
@@ -1,6 +1,6 @@
 package com.widyu.domain.auth.repository;
 
-import com.widyu.domain.auth.domain.TemporaryMember;
+import com.widyu.domain.auth.entity.TemporaryMember;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
