+++ src/main/java/com/widyu/domain/auth/repository/VerificationCodeRepository.java
@@ -1,6 +1,6 @@
 package com.widyu.domain.auth.repository;
 
-import com.widyu.domain.auth.domain.VerificationCode;
+import com.widyu.domain.auth.entity.VerificationCode;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
